requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.3
openai==1.51.0
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.4